+++ pkg/apis/tooling/v1alpha1/azure_tool_types.go
@@ -1,52 +1,18 @@
 package v1alpha1
 
 import (
-	apiextensionsv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
+	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/yaml"
+
+	"github.com/giantswarm/apiextensions/pkg/crd"
 )
 
-const azureToolCRDYAML = `
-apiVersion: apiextensions.k8s.io/v1beta1
-kind: CustomResourceDefinition
-metadata:
-  name: azuretools.tooling.giantswarm.io
-spec:
-  group: tooling.giantswarm.io
-  scope: Namespaced
-  version: v1alpha1
-  names:
-    kind: AzureTool
-    plural: azuretools
-    singular: azuretool
-  validation:
-    openAPIV3Schema:
-      properties:
-        spec:
-          type: object
-          properties:
-            workspace:
-              type: object
-              properties:
-                id:
-                  type: string
-                mode:
-                  type: string
-          required:
-          - workspace
-`
+const (
+	kindAzureTool = "AzureTool"
+)
 
-var azureToolCRD *apiextensionsv1beta1.CustomResourceDefinition
-
-func init() {
-	err := yaml.Unmarshal([]byte(azureToolCRDYAML), &azureToolCRD)
-	if err != nil {
-		panic(err)
-	}
-}
-
-func NewAzureToolCRD() *apiextensionsv1beta1.CustomResourceDefinition {
-	return azureToolCRD.DeepCopy()
+func NewAzureToolCRD() *v1beta1.CustomResourceDefinition {
+	return crd.LoadV1Beta1(group, kindAzureTool)
 }
 
 // +genclient
@@ -59,14 +25,14 @@ type AzureTool struct {
 }
 
 type AzureToolSpec struct {
-	// Workspace refers to the Azure Log Analytics Workspace
+	// Workspace refers to the Azure Log Analytics Workspace.
 	Workspace AzureToolWorkspace `json:"workspace" yaml:"workspace"`
 }
 
 type AzureToolWorkspace struct {
-	// ID is the Workspace ID
+	// ID is the Workspace ID.
 	ID string `json:"id" yaml:"id"`
-	// Mode is the Workspace access control mode
+	// Mode is the mode that the Workspace is running in.
 	Mode string `json:"mode" yaml:"mode"`
 }